moving common fixtures to conftest
